+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Fixed
+ - Resolve brittle handling of class- and module-level assignments in flake8 plugin
 
 
 ## [2.1.0] — 2024-03-08